+++ src/components/App/App.js
@@ -81,10 +81,11 @@ function App() {
     localStorage.removeItem('padachone_msg');
     localStorage.removeItem('padachone_msg1');
     localStorage.removeItem('padachone_msg2');
-    if (!localStorage.getItem('padachone_msg3')) {
-      const message = `Would you like an Email Subscribe option? - admirer@padachone.com`;
+    localStorage.removeItem('padachone_msg3');
+    if (!localStorage.getItem('padachone_msg4')) {
+      const message = `Email Subscription option is available now! Check homepage footer.`;
       setMsg(() => {
-        localStorage.setItem('padachone_msg3', message)
+        localStorage.setItem('padachone_msg4', message)
         return [true, message]
       });      
     }